Sidebar navigation

+++ src/routes/account/+page.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
 	import { goto } from '$app/navigation';
-	import AppFooter from '$lib/components/AppFooter.svelte';
 	import AppHeader from '$lib/components/AppHeader.svelte';
 	import debounce from '$lib/debounce';
 	import { page } from '$app/state';
@@ -12,6 +11,7 @@
 	import * as AlertDialog from '$lib/components/ui/alert-dialog';
 	import { Err } from '$domain/errors';
 	import UserSettings from '$lib/user-settings/UserSettings.svelte';
+	import { get } from 'svelte/store';
 
 	let taskCount = $state<number>(0);
 	let isDeleting = $state(false);
@@ -21,42 +21,30 @@
 
 	const debouncedUpdateUser = debounce(authAPI.updateUser, 200);
 
-	async function loadTaskCount() {
-		if ($authState.status !== 'signed-in') return;
-
-		const [userTasks, error] = await tasksAPI.getAllUserTasks({ userId: $authState.user.id });
-		if (userTasks) {
-			taskCount = userTasks.length;
-		} else {
-			taskCount = 0;
-			Err.UNHANDLED(error, 'Failed to load task count:');
-		}
-	}
-
 	async function handleDeleteUser() {
 		if ($authState.status !== 'signed-in' || isDeleting) return;
 
 		isDeleting = true;
 		try {
-			const [roots, getRootsError] = await tasksAPI.getRootTasks();
-			if (getRootsError) Err.UNHANDLED(getRootsError);
+			tasksAPI.getAllUserTasks({ userId: $authState.user.id }).subscribe(async (allTasks) => {
+				if (allTasks.status === 'error') {
+					Err.UNHANDLED(allTasks.error);
+				} else if (allTasks.status === 'resolved') {
+					await tasksAPI.deleteTasks({
+						ids: allTasks.data.map((r) => r.id)
+					});
 
-			await tasksAPI.deleteTasks({
-				ids: roots.map((r) => r.id)
+					// Delete the user account
+					const [_, deleteUserError] = await authAPI.deleteUser({ userId: $authState.user.id });
+					if (deleteUserError) Err.UNHANDLED(deleteUserError);
+				}
 			});
-
-			// Delete the user account
-			const [_, deleteUserError] = await authAPI.deleteUser({ userId: $authState.user.id });
-			if (deleteUserError) Err.UNHANDLED(deleteUserError);
 
 			// TODO:Temp anonymous accounts disabled
 			/* const defaultUserResult = await auth.getDefaultUser();
 			if (defaultUserResult.isOk()) {
 				await auth.switchUser(defaultUserResult.value.id);
 			} */
-
-			// Redirect to login page since current user is deleted
-			// goto('/login');
 		} catch (error) {
 			isDeleting = false;
 			Err.UNHANDLED(error, 'Failed to delete user:');
@@ -196,6 +184,5 @@
 				<UserSettings />
 			</div>
 		</div>
-		<AppFooter />
 	</div>
 {/if}
